Initial commit
This branch is an example of how we might alter the structure of the
package. In my opinion, this makes the purpose and function of the
package as a make-library to turn raw data into a processed file a bit
clearer.

The major changes are:
        Variable scripts are made into functions, and live in separate
        files in the R/ directory.

        Making PRIOGrid is done by calling the make_pg function, which
        calls each separate variable function. The variable function is
        given a directory by the make_pg function, containing the
        constituent data, and writes finished data to a temporary
        folder. At the end of the function, it stitches all of these
        intermediate files together, forming the "final" PRIOGrid.

        "Helper" functions, such as our super-smart and fast functions
        for doing intersects (forthcoming) live in other files under
        R/.

This is just a suggestion, but i am pretty sure that this is the best
way to organize this package. Looking forward to your feedback!

Peder